feat(react-email): Substitute implementation of `render` with the one from `renderAsync`
