+++ packages/render/src/index.ts
@@ -1,5 +1,3 @@
 export * from "./render";
-export * from "./render-async";
-
 export * from "./options";
 export * from "./plain-text-selectors";
