+++ packages/render/src/render-web.spec.tsx
@@ -4,9 +4,9 @@
 
 import { Template } from "./utils/template";
 import { Preview } from "./utils/preview";
-import { renderAsync } from "./render-async";
+import { render } from "./render";
 
-describe("renderAsync on the browser environment", () => {
+describe("render on the browser environment", () => {
   beforeEach(() => {
     vi.mock(
       "react-dom/server",
@@ -44,7 +44,7 @@ describe("renderAsync on the browser environment", () => {
       };
     });
 
-    const actualOutput = await renderAsync(<Template firstName="Jim" />);
+    const actualOutput = await render(<Template firstName="Jim" />);
 
     expect(actualOutput).toMatchInlineSnapshot(
       '"<!DOCTYPE html PUBLIC \\"-//W3C//DTD XHTML 1.0 Transitional//EN\\" \\"http://www.w3.org/TR/xhtml1/DTD/xhtml1-transitional.dtd\\"><h1>Welcome, <!-- -->Jim<!-- -->!</h1><img alt=\\"test\\" src=\\"img/test.png\\"/><p>Thanks for trying our product. We&#x27;re thrilled to have you on board!</p>"',
@@ -58,7 +58,7 @@ describe("renderAsync on the browser environment", () => {
   });
 
   it("converts a React component into HTML", async () => {
-    const actualOutput = await renderAsync(<Template firstName="Jim" />);
+    const actualOutput = await render(<Template firstName="Jim" />);
 
     expect(actualOutput).toMatchInlineSnapshot(
       '"<!DOCTYPE html PUBLIC \\"-//W3C//DTD XHTML 1.0 Transitional//EN\\" \\"http://www.w3.org/TR/xhtml1/DTD/xhtml1-transitional.dtd\\"><h1>Welcome, <!-- -->Jim<!-- -->!</h1><img alt=\\"test\\" src=\\"img/test.png\\"/><p>Thanks for trying our product. We&#x27;re thrilled to have you on board!</p>"',
@@ -66,7 +66,7 @@ describe("renderAsync on the browser environment", () => {
   });
 
   it("converts a React component into PlainText", async () => {
-    const actualOutput = await renderAsync(<Template firstName="Jim" />, {
+    const actualOutput = await render(<Template firstName="Jim" />, {
       plainText: true,
     });
 
@@ -78,7 +78,7 @@ describe("renderAsync on the browser environment", () => {
   });
 
   it("converts to plain text and removes reserved ID", async () => {
-    const actualOutput = await renderAsync(<Preview />, {
+    const actualOutput = await render(<Preview />, {
       plainText: true,
     });
